+++ packages/alice/lib/core/alice_core.dart
@@ -1,11 +1,11 @@
 import 'dart:async' show FutureOr, StreamSubscription;
 
-import 'package:alice/core/alice_logger.dart';
 import 'package:alice/core/alice_storage.dart';
 import 'package:alice/core/alice_utils.dart';
 import 'package:alice/helper/alice_export_helper.dart';
 import 'package:alice/core/alice_notification.dart';
 import 'package:alice/helper/operating_system.dart';
+import 'package:alice/model/alice_configuration.dart';
 import 'package:alice/model/alice_export_result.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_http_error.dart';
@@ -15,37 +15,9 @@ import 'package:alice/ui/common/alice_navigation.dart';
 import 'package:alice/utils/shake_detector.dart';
 import 'package:flutter/material.dart';
 
-typedef AliceOnCallsChanged = Future<void> Function(List<AliceHttpCall>? calls);
-
 class AliceCore {
-  /// Should user be notified with notification if there's new request caught
-  /// by Alice
-  final bool showNotification;
-
-  /// Should inspector be opened on device shake (works only with physical
-  /// with sensors)
-  final bool showInspectorOnShake;
-
-  /// Icon url for notification
-  final String notificationIcon;
-
-  /// Storage used for Alice to keep calls data.
-  final AliceStorage _aliceStorage;
-
-  /// Logger used for Alice to keep logs;
-  final AliceLogger _aliceLogger;
-
-  ///Directionality of app. If null then directionality of context will be used.
-  final TextDirection? directionality;
-
-  ///Flag used to show/hide share button
-  final bool? showShareButton;
-
-  /// Navigator key used for inspector navigator.
-  GlobalKey<NavigatorState>? navigatorKey;
-
-  /// Flag used to determine whether is inspector opened
-  bool _isInspectorOpened = false;
+  /// Configuration of Alice
+  late AliceConfiguration _configuration;
 
   /// Detector used to detect device shakes
   ShakeDetector? _shakeDetector;
@@ -56,27 +28,21 @@ class AliceCore {
   /// Subscription for call changes
   StreamSubscription<List<AliceHttpCall>>? _callsSubscription;
 
+  /// Flag used to determine whether is inspector opened
+  bool _isInspectorOpened = false;
+
   /// Creates alice core instance
-  AliceCore(
-    this.navigatorKey, {
-    required this.showNotification,
-    required this.showInspectorOnShake,
-    required this.notificationIcon,
-    required AliceStorage aliceStorage,
-    required AliceLogger aliceLogger,
-    this.directionality,
-    this.showShareButton,
-  })  : _aliceStorage = aliceStorage,
-        _aliceLogger = aliceLogger {
+  AliceCore({required AliceConfiguration configuration}) {
+    _configuration = configuration;
     _subscribeToCallChanges();
-    if (showNotification) {
+    if (_configuration.showNotification) {
       _notification = AliceNotification();
       _notification?.configure(
-        notificationIcon: notificationIcon,
+        notificationIcon: _configuration.notificationIcon,
         openInspectorCallback: navigateToCallListScreen,
       );
     }
-    if (showInspectorOnShake) {
+    if (_configuration.showInspectorOnShake) {
       if (OperatingSystem.isAndroid || OperatingSystem.isMacOS) {
         _shakeDetector = ShakeDetector.autoStart(
           onPhoneShake: navigateToCallListScreen,
@@ -84,6 +50,14 @@ class AliceCore {
         );
       }
     }
+  }
+
+  /// Returns current configuration
+  AliceConfiguration get configuration => _configuration;
+
+  /// Set custom navigation key. This will help if there's route library.
+  void setNavigatorKey(GlobalKey<NavigatorState> navigatorKey) {
+    _configuration = _configuration.copyWith(navigatorKey: navigatorKey);
   }
 
   /// Dispose subjects and subscriptions
@@ -95,7 +69,7 @@ class AliceCore {
   /// Called when calls has been updated
   Future<void> _onCallsChanged(List<AliceHttpCall>? calls) async {
     if (calls != null && calls.isNotEmpty) {
-      final AliceStats stats = _aliceStorage.getStats();
+      final AliceStats stats = _configuration.aliceStorage.getStats();
       _notification?.showStatsNotification(
         context: getContext()!,
         stats: stats,
@@ -105,7 +79,7 @@ class AliceCore {
 
   /// Opens Http calls inspector. This will navigate user to the new fullscreen
   /// page where all listened http calls can be viewed.
-  void navigateToCallListScreen() {
+  Future<void> navigateToCallListScreen() async {
     final BuildContext? context = getContext();
     if (context == null) {
       AliceUtils.log(
@@ -116,56 +90,60 @@ class AliceCore {
     }
     if (!_isInspectorOpened) {
       _isInspectorOpened = true;
-      AliceNavigation.navigateToCallsList(core: this, logger: _aliceLogger)
-          .then((_) => _isInspectorOpened = false);
+      await AliceNavigation.navigateToCallsList(core: this);
+      _isInspectorOpened = false;
     }
   }
 
   /// Get context from navigator key. Used to open inspector route.
-  BuildContext? getContext() => navigatorKey?.currentState?.overlay?.context;
+  BuildContext? getContext() =>
+      _configuration.navigatorKey?.currentState?.overlay?.context;
 
   /// Add alice http call to calls subject
-  FutureOr<void> addCall(AliceHttpCall call) => _aliceStorage.addCall(call);
+  FutureOr<void> addCall(AliceHttpCall call) =>
+      _configuration.aliceStorage.addCall(call);
 
   /// Add error to existing alice http call
   FutureOr<void> addError(AliceHttpError error, int requestId) =>
-      _aliceStorage.addError(error, requestId);
+      _configuration.aliceStorage.addError(error, requestId);
 
   /// Add response to existing alice http call
   FutureOr<void> addResponse(AliceHttpResponse response, int requestId) =>
-      _aliceStorage.addResponse(response, requestId);
+      _configuration.aliceStorage.addResponse(response, requestId);
 
   /// Remove all calls from calls subject
-  FutureOr<void> removeCalls() => _aliceStorage.removeCalls();
+  FutureOr<void> removeCalls() => _configuration.aliceStorage.removeCalls();
 
   /// Selects call with given [requestId]. It may return null.
   @protected
   AliceHttpCall? selectCall(int requestId) =>
-      _aliceStorage.selectCall(requestId);
+      _configuration.aliceStorage.selectCall(requestId);
 
   /// Returns stream which returns list of HTTP calls
-  Stream<List<AliceHttpCall>> get callsStream => _aliceStorage.callsStream;
+  Stream<List<AliceHttpCall>> get callsStream =>
+      _configuration.aliceStorage.callsStream;
 
   /// Returns all stored HTTP calls.
-  List<AliceHttpCall> getCalls() => _aliceStorage.getCalls();
+  List<AliceHttpCall> getCalls() => _configuration.aliceStorage.getCalls();
 
   /// Save all calls to file.
-  Future<AliceExportResult> saveCallsToFile(BuildContext context) {
-    return AliceExportHelper.saveCallsToFile(context, _aliceStorage.getCalls());
-  }
+  Future<AliceExportResult> saveCallsToFile(BuildContext context) =>
+      AliceExportHelper.saveCallsToFile(
+          context, _configuration.aliceStorage.getCalls());
 
   /// Adds new log to Alice logger.
-  void addLog(AliceLog log) => _aliceLogger.add(log);
+  void addLog(AliceLog log) => _configuration.aliceLogger.add(log);
 
   /// Adds list of logs to Alice logger
-  void addLogs(List<AliceLog> logs) => _aliceLogger.addAll(logs);
+  void addLogs(List<AliceLog> logs) => _configuration.aliceLogger.addAll(logs);
 
   /// Returns flag which determines whether inspector is opened
   bool get isInspectorOpened => _isInspectorOpened;
 
   /// Subscribes to storage for call changes.
   void _subscribeToCallChanges() {
-    _callsSubscription = _aliceStorage.callsStream.listen(_onCallsChanged);
+    _callsSubscription =
+        _configuration.aliceStorage.callsStream.listen(_onCallsChanged);
   }
 
   /// Unsubscribes storage for call changes.