+++ packages/alice/lib/ui/call_details/page/alice_call_details_page.dart
@@ -74,17 +74,18 @@ class _AliceCallDetailsPageState extends State<AliceCallDetailsPage>
                       AliceCallErrorScreen(call: widget.call),
                     ],
                   ),
-                  floatingActionButton: widget.core.showShareButton ?? false
-                      ? FloatingActionButton(
-                          backgroundColor: AliceTheme.lightRed,
-                          key: const Key('share_key'),
-                          onPressed: _shareCall,
-                          child: const Icon(
-                            Icons.share,
-                            color: AliceTheme.white,
-                          ),
-                        )
-                      : null,
+                  floatingActionButton:
+                      widget.core.configuration.showShareButton
+                          ? FloatingActionButton(
+                              backgroundColor: AliceTheme.lightRed,
+                              key: const Key('share_key'),
+                              onPressed: _shareCall,
+                              child: const Icon(
+                                Icons.share,
+                                color: AliceTheme.white,
+                              ),
+                            )
+                          : null,
                 ),
               );
             }
@@ -102,6 +103,8 @@ class _AliceCallDetailsPageState extends State<AliceCallDetailsPage>
     );
   }
 
+  /// Called when share button has been pressed. It encodes the [widget.call]
+  /// and tries to invoke system action to share it.
   void _shareCall() async {
     await AliceExportHelper.shareCall(context: context, call: widget.call);
   }