+++ packages/alice/lib/ui/calls_list/widget/alice_calls_list_screen.dart
@@ -19,6 +19,7 @@ class AliceCallsListScreen extends StatelessWidget {
   final bool sortAscending;
   final void Function(AliceHttpCall) onListItemClicked;
 
+  /// Returns sorted calls based [sortOption] and [sortAscending].
   List<AliceHttpCall> get _sortedCalls => switch (sortOption) {
         AliceCallsListSortOption.time => sortAscending
             ? (calls