+++ packages/alice/lib/ui/calls_list/widget/alice_inspector_screen.dart
@@ -5,6 +5,8 @@ import 'package:alice/ui/calls_list/widget/alice_calls_list_screen.dart';
 import 'package:alice/ui/calls_list/widget/alice_empty_logs_widget.dart';
 import 'package:flutter/material.dart';
 
+/// Screen which is hosted in calls list page. It displays HTTP calls. It allows
+/// to search call and sort items based on provided criteria.
 class AliceInspectorScreen extends StatefulWidget {
   const AliceInspectorScreen({
     super.key,