+++ packages/alice/lib/ui/calls_list/widget/alice_log_list_widget.dart
@@ -2,6 +2,8 @@ import 'dart:convert';
 
 import 'package:alice/model/alice_log.dart';
 import 'package:alice/model/alice_translation.dart';
+import 'package:alice/ui/calls_list/widget/alice_empty_logs_widget.dart';
+import 'package:alice/ui/calls_list/widget/alice_error_logs_widget.dart';
 import 'package:alice/ui/common/alice_context_ext.dart';
 import 'package:alice/ui/common/alice_scroll_behavior.dart';
 import 'package:alice/ui/common/alice_theme.dart';
@@ -12,15 +14,13 @@ import 'package:flutter/services.dart';
 /// Widget which renders log list for calls list page.
 class AliceLogListWidget extends StatefulWidget {
   const AliceLogListWidget({
-    required this.logsListenable,
+    required this.logsStream,
     required this.scrollController,
-    required this.emptyWidget,
     super.key,
   });
 
-  final ValueListenable<List<AliceLog>> logsListenable;
+  final Stream<List<AliceLog>>? logsStream;
   final ScrollController? scrollController;
-  final Widget emptyWidget;
 
   @override
   State<AliceLogListWidget> createState() => _AliceLogListWidgetState();
@@ -32,11 +32,23 @@ class _AliceLogListWidgetState extends State<AliceLogListWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return ValueListenableBuilder<List<AliceLog>>(
-      valueListenable: widget.logsListenable,
-      builder: (_, List<AliceLog> logs, __) {
+    return StreamBuilder<List<AliceLog>>(
+      stream: widget.logsStream,
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.none ||
+            snapshot.connectionState == ConnectionState.waiting) {
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        }
+
+        if (snapshot.hasError) {
+          return const AliceErrorLogsWidget();
+        }
+
+        final logs = snapshot.data ?? [];
         if (logs.isEmpty) {
-          return widget.emptyWidget;
+          return const AliceEmptyLogsWidget();
         }
 
         final List<AliceLog> filteredLogs = [