+++ packages/alice/test/core/alice_memory_storage_test.dart
@@ -6,7 +6,7 @@ import 'package:alice/model/alice_http_response.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
 
-import '../mocked_data.dart';
+import '../mock/mocked_data.dart';
 
 void main() {
   late AliceMemoryStorage storage;
@@ -39,12 +39,14 @@ void main() {
       storage
           .addCall(MockedData.getHttpCallWithResponseStatus(statusCode: 500));
       storage.addCall(MockedData.getLoadingHttpCall());
+      storage.addCall(MockedData.getHttpCallWithResponseStatus(statusCode: -1));
+      storage.addCall(MockedData.getHttpCallWithResponseStatus(statusCode: 0));
 
       expect(storage.getStats(), (
-        total: 8,
+        total: 10,
         successes: 2,
         redirects: 2,
-        errors: 3,
+        errors: 5,
         loading: 1,
       ));
     });