+++ packages/alice/CHANGELOG.md
@@ -1,4 +1,15 @@
+# 1.0.0-dev.10
+
+* [BREAKING_CHANGE] Removed `maxCallsCount`. To change call count, use `storage` constructor parameter
+  with call count.
+* [BREAKING_CHANGE] Replaced configuration parameter of Alice with `AliceConfiguration`.
+* Fixed issue with invalid count of calls in notification.
+* Added payload to notification. 
+* General refactor of code base.
+* Updated documentation.
+
 # 1.0.0-dev.9
+
 * Fixed saving issue with Android 13 onwards.
 * Added unit tests.
 * Updated CI/CD task for tests.
@@ -6,11 +17,13 @@
 * Updated metadata.
 
 # 1.0.0-dev.8
+
 * Added storage abstractions (by Klemen Tusar https://github.com/techouse).
 * Added in memory storage implementation (by Klemen Tusar https://github.com/techouse).
 * Added translations.
 
 # 1.0.0-dev.7
+
 * Refactored UI code.
 
 # 1.0.0-dev.6